Cube vertices revision and there is issue at displaying image on imgui when using directx11 (what is problem..)

+++ SandBox/src/RenderGraph/Pass/PBRShaderPass.h
@@ -55,11 +55,11 @@ namespace QCat
 			float roughness;
 			float ambientocclusion;
 
-			int IsAlbedoMap;
-			int IsNormalMap;
-			int IsMetallicMap;
-			int IsRoughnessMap;
-			int IsAoMap;
+			float IsAlbedoMap;
+			float IsNormalMap;
+			float IsMetallicMap;
+			float IsRoughnessMap;
+			float IsAoMap;
 		};
 		struct light
 		{
